Extract simulations to a useSimulation hook
